Move unit tests to corresponding implementation package
changelog_begin
changelog_end
